requirements.txt
vosk>=0.3.45
pyttsx3>=2.90
pyautogui>=0.9.54
pywinauto>=0.6.8
sounddevice>=0.4.6
numpy>=1.24.0
requests>=2.31.0
mediapipe>=0.10.9
opencv-python>=4.9.0
pystray>=0.19.5
Pillow>=10.0.0
